Move kafka consumer out
